added view action on history page, refactored Analysis component to reduce line count
