+++ frontend/src/components/Analysis/fiatLabels.js
@@ -1,0 +1,43 @@
+export const fiatLabels = {
+	AED: "United Arab Emirates Dirham",
+	ARS: "Argentine Peso",
+	AUD: "Australian Dollar",
+	BDT: "Bangladeshi Taka",
+	BRL: "Brazilian Real",
+	CAD: "Canadian Dollar",
+	CHF: "Swiss Franc",
+	CLP: "Chilean Peso",
+	CNY: "Chinese Yuan",
+	CZK: "Czech Koruna",
+	DKK: "Danish Krone",
+	EUR: "Euro",
+	GBP: "Pound sterling",
+	HKD: "Hong Kong Dollar",
+	HUF: "Hungarian Forint",
+	IDR: "Indonesian Rupiah",
+	IKR: "Icelandic Króna",
+	ILS: "Israeli New Shekel",
+	INR: "Indian Rupee",
+	JPY: "Japanese Yen",
+	KRW: "South Korean won",
+	MMK: "Myanmar Kyat",
+	MXN: "Mexican Peso",
+	MYR: "Malaysian Ringgit",
+	NGN: "Nigerian Naira",
+	NOK: "Norwegian krone",
+	NZD: "New Zealand Dollar",
+	PHP: "Philippine peso",
+	PKR: "Pakistani Rupee",
+	PLN: "Poland złoty",
+	RUB: "Russian Ruble",
+	SEK: "Swedish Krona",
+	SGD: "Singapore Dollar",
+	THB: "Thai Baht",
+	TRY: "Turkish lira",
+	TWD: "New Taiwan dollar",
+	UAH: "Ukrainian hryvnia",
+	USD: "United States Dollar",
+	VEB: "Venezuelan bolivar",
+	VND: "Vietnamese dong",
+	ZAR: "South African Rand",
+};
