+++ frontend/src/components/Landing/Landing.css
@@ -37,7 +37,7 @@
 	font-size: 5em;
 	font-family: "Source Sans Pro", sans-serif;
 	font-weight: 600;
-	color: #4b4b4b;
+	color: #2e2e2e;
 }
 
 .name-one-liner {
@@ -61,7 +61,6 @@ p {
 	display: flex;
 	justify-content: center;
 	margin-top: 25px;
-
 }
 
 .password {
